Fix SFLP conversion

+++ MIDAS/lib/lsm6dsv320x-pid/lsm6dsv320x.h
@@ -526,7 +526,7 @@ public:
     float from_fs64_to_mg(int16_t lsb);
     float from_fs2000_to_mdps(int16_t lsb);
 
-    uint32_t half_to_float(uint16_t h);
+    float half_to_float(uint16_t h);
   
     void get_lowg_acceleration_from_fs8_to_g(float *ax, float *ay, float *az);
     void get_highg_acceleration_from_fs64_to_g(float *ax, float *ay, float *az);
